+++ src/main/java/com/huotu/huobanmall/model/app/AppSalesReportListModel.java
@@ -24,7 +24,7 @@ public class AppSalesReportListModel {
     /**
      * 金额
      */
-    private float money;
+    private double money;
 
     public Date getTime() {
         return time;
@@ -34,11 +34,11 @@ public class AppSalesReportListModel {
         this.time = time;
     }
 
-    public float getMoney() {
+    public double getMoney() {
         return money;
     }
 
-    public void setMoney(float money) {
+    public void setMoney(double money) {
         this.money = money;
     }
 }
